+++ src/us/kbase/workspace/kbase/WorkspaceAdministration.java
@@ -31,7 +31,9 @@ import us.kbase.workspace.GetObjects2Params;
 import us.kbase.workspace.GetPermissionsMassParams;
 import us.kbase.workspace.GrantModuleOwnershipParams;
 import us.kbase.workspace.ListObjectsParams;
+import us.kbase.workspace.ListWorkspaceIDsParams;
 import us.kbase.workspace.ListWorkspaceInfoParams;
+import us.kbase.workspace.ObjectIdentity;
 import us.kbase.workspace.RemoveModuleOwnershipParams;
 import us.kbase.workspace.SaveObjectsParams;
 import us.kbase.workspace.SetGlobalPermissionsParams;
@@ -63,9 +65,11 @@ public class WorkspaceAdministration {
 	private static final String REMOVE_MODULE_OWNERSHIP = "removeModuleOwnership";
 	private static final String GRANT_MODULE_OWNERSHIP = "grantModuleOwnership";
 	private static final String LIST_WORKSPACES = "listWorkspaces";
+	private static final String LIST_WORKSPACE_IDS = "listWorkspaceIDs";
 	private static final String LIST_OBJECTS = "listObjects";
 	private static final String SAVE_OBJECTS = "saveObjects";
 	private static final String GET_OBJECT_INFO = "getObjectInfo";
+	private static final String GET_OBJECT_HIST = "getObjectHistory";
 	private static final String GET_OBJECTS = "getObjects";
 	private static final String SET_GLOBAL_PERMISSION = "setGlobalPermission";
 	private static final String GET_WORKSPACE_INFO = "getWorkspaceInfo";
@@ -233,6 +237,12 @@ public class WorkspaceAdministration {
 			getLogger().info(GET_OBJECT_INFO);
 			return wsmeth.getObjectInformation(params, new WorkspaceUser(putativeAdmin), true);
 		}
+		if (GET_OBJECT_HIST.equals(fn)) {
+			final ObjectIdentity params = getParams(cmd, ObjectIdentity.class);
+			// method has its own logging
+			getLogger().info(GET_OBJECT_HIST);
+			return wsmeth.getObjectHistory(params, new WorkspaceUser(putativeAdmin), true);
+		}
 		if (GET_OBJECTS.equals(fn)) {
 			final GetObjects2Params params = getParams(cmd, GetObjects2Params.class);
 			// method has its own logging
@@ -245,6 +255,12 @@ public class WorkspaceAdministration {
 			final WorkspaceUser user = getUser(cmd, token);
 			getLogger().info(LIST_WORKSPACES + " " + user.getUser());
 			return wsmeth.listWorkspaceInfo(params, user);
+		}
+		if (LIST_WORKSPACE_IDS.equals(fn)) {
+			final ListWorkspaceIDsParams params = getParams(cmd, ListWorkspaceIDsParams.class);
+			final WorkspaceUser user = getUser(cmd, token);
+			getLogger().info(LIST_WORKSPACE_IDS + " " + user.getUser());
+			return wsmeth.listWorkspaceIDs(params, user);
 		}
 		if (LIST_OBJECTS.equals(fn)) {
 			final ListObjectsParams params = getParams(cmd, ListObjectsParams.class);