+++ src/us/kbase/workspace/kbase/WorkspaceServerMethods.java
@@ -11,6 +11,7 @@ import static us.kbase.workspace.kbase.ArgUtils.translateObjectData;
 import static us.kbase.workspace.kbase.ArgUtils.longToBoolean;
 import static us.kbase.workspace.kbase.ArgUtils.longToInt;
 import static us.kbase.workspace.kbase.ArgUtils.objInfoToTuple;
+import static us.kbase.workspace.kbase.IdentifierUtils.processObjectIdentifier;
 import static us.kbase.workspace.kbase.IdentifierUtils.processObjectSpecifications;
 import static us.kbase.workspace.kbase.IdentifierUtils.processWorkspaceIdentifier;
 import static us.kbase.workspace.kbase.KBasePermissions.translatePermission;
@@ -50,7 +51,10 @@ import us.kbase.workspace.GetObjects2Params;
 import us.kbase.workspace.GetObjects2Results;
 import us.kbase.workspace.GrantModuleOwnershipParams;
 import us.kbase.workspace.ListObjectsParams;
+import us.kbase.workspace.ListWorkspaceIDsParams;
+import us.kbase.workspace.ListWorkspaceIDsResults;
 import us.kbase.workspace.ListWorkspaceInfoParams;
+import us.kbase.workspace.ObjectIdentity;
 import us.kbase.workspace.ObjectSaveData;
 import us.kbase.workspace.RemoveModuleOwnershipParams;
 import us.kbase.workspace.SaveObjectsParams;
@@ -67,6 +71,7 @@ import us.kbase.workspace.database.Permission;
 import us.kbase.workspace.database.Provenance;
 import us.kbase.workspace.database.Types;
 import us.kbase.workspace.database.User;
+import us.kbase.workspace.database.UserWorkspaceIDs;
 import us.kbase.workspace.database.Workspace;
 import us.kbase.workspace.database.WorkspaceIdentifier;
 import us.kbase.workspace.database.WorkspaceInformation;
@@ -507,6 +512,33 @@ public class WorkspaceServerMethods {
 				longToBoolean(params.getShowOnlyDeleted())));
 	}
 	
+	/** Lists IDs of workspaces to which the user has access. Should be faster than
+	 * {@link #listWorkspaceInfo(ListWorkspaceInfoParams, WorkspaceUser)}.
+	 * @param params the method parameters.
+	 * @param user the user, or null if anonymous.
+	 * @return the results of the method.
+	 * @throws CorruptWorkspaceDBException if corrupt data was found in the storage system.
+	 * @throws WorkspaceCommunicationException if a communication error occurred with the storage
+	 * system.
+	 */
+	public ListWorkspaceIDsResults listWorkspaceIDs(
+			final ListWorkspaceIDsParams params,
+			WorkspaceUser user)
+			throws WorkspaceCommunicationException, CorruptWorkspaceDBException {
+		checkAddlArgs(params.getAdditionalProperties(), params.getClass());
+		if (longToBoolean(params.getOnlyGlobal())) {
+			user = null;
+			params.setExcludeGlobal(0L);
+		}
+		final Permission p = params.getPerm() == null ? null :
+			translatePermission(params.getPerm());
+		final UserWorkspaceIDs wsids = ws.listWorkspaceIDs(
+				user, p, longToBoolean(params.getExcludeGlobal(), true));
+		return new ListWorkspaceIDsResults()
+				.withWorkspaces(new LinkedList<>(wsids.getWorkspaceIDs()))
+				.withPub(new LinkedList<>(wsids.getPublicWorkspaceIDs()));
+	}
+	
 	/** List objects in one or more workspaces.
 	 * @param params the parameters determining which workspace objects will be listed.
 	 * @param user the user listing the objects, or null for an anonymous user.
@@ -603,5 +635,27 @@ public class WorkspaceServerMethods {
 			}
 		}
 		return lop;
+	}
+
+	/** Get all versions of an object.
+	 * @param object the object.
+	 * @param user the user making the request.
+	 * @param asAdmin true to ignore the user and request as an admin.
+	 * @return the object versions.
+	 * @throws InaccessibleObjectException if the object is inaccessible.
+	 * @throws NoSuchObjectException if there is no such object.
+	 * @throws WorkspaceCommunicationException if a communication error occurs when contacting the
+	 * storage system.
+	 * @throws CorruptWorkspaceDBException if corrupt data is found in the storage system.
+	 */
+	public List<Tuple11<Long, String, String, String, Long, String, Long, String, String,
+			Long, Map<String, String>>> getObjectHistory(
+			final ObjectIdentity object,
+			final WorkspaceUser user,
+			final boolean asAdmin)
+			throws WorkspaceCommunicationException, InaccessibleObjectException,
+			CorruptWorkspaceDBException, NoSuchObjectException {
+		final ObjectIdentifier oi = processObjectIdentifier(object);
+		return objInfoToTuple(ws.getObjectHistory(user, oi, asAdmin), true);
 	}
 }